src/requirements.txt
keras
matplotlib
numpy
pandas
plotly
python-binance
requests
statsmodels